feat: Added some Custom hooks.
